pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.1.17",
    "h11>=0.16.0",
    "pydantic-core>=2.35.1",
    "starlette>=0.47.1",
    "typing-inspection>=0.4.1",
]
